chore: reactor local test into runner and suit
